Cleanup of rxjs; removing cruft we no longer use (most was associated with the 'beta' history implementation)
