--- client/src/utils/observable/monitorBackboneModel.js
@@ -1,9 +1,0 @@
-import { fromEvent } from "rxjs";
-import { map } from "rxjs/operators";
-
-// prettier-ignore
-export function monitorBackboneModel(sourceModel, evtName = "change") {
-    return fromEvent(sourceModel, evtName).pipe(
-        map(([model]) => model.toJSON())
-    );
-}
